White space issue fixed
Before this edit, on lines 49 and 62, you were using `createText` to insert a whitespace before and after certain elements. This would add the whitespace so the sentence didn't have any missing whitespaces, but those new whitespaces weren't being added to the allElements array. This makes those newly create whitespaces to be un-splittable. By taking line 54 and replacing lines 46. and 62. you add these whitespaces to the allElements array, allowing them to be split correctly.